backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
python-dotenv
httpx
pydantic